+++ packages/core/src/utils/shaderCodeUtils.ts
@@ -1,60 +1,11 @@
 import type { Effect } from "@babylonjs/core/Materials/effect";
+import type { ShaderCode } from "./buildTools/shaderCode.types";
 
 /**
  * The shader code decorator.
  * Used to decorate the names of uniform, function and const variables for easier parsing.
  */
 export const decorateChar = "_";
-
-/**
- * Describes a shader function.
- */
-export type ShaderFunction = {
-    /**
-     * The name of the function.
-     */
-    name: string;
-
-    /**
-     * The code of the function.
-     */
-    code: string;
-};
-
-/**
- * Describes a shader code.
- */
-export type ShaderCode = {
-    /**
-     * The declaration of the const variables.
-     */
-    const?: string;
-
-    /**
-     * The declaration of the uniform variables.
-     */
-    uniform?: string;
-
-    /**
-     * The declaration of the uniform variables that should be common for all ShaderBlock instances using this shader code.
-     */
-    uniformSingle?: string;
-
-    /**
-     * The name of the main function.
-     */
-    mainFunctionName: string;
-
-    /**
-     * The name of the input texture which is passed through if the block is disabled.
-     */
-    mainInputTexture?: string;
-
-    /**
-     * The list of functions used in the shader.
-     */
-    functions: ShaderFunction[];
-};
 
 /**
  * Describes a shader program.
